### 2025-10-24 #9120 @crvernon
@taylanot - Thank you for your submission to JOSS.  We are handling a large backlog of submissions and we will get you a topic editor assigned as soon as one becomes available.
