+++ docs/paper/ref.bib
@@ -96,6 +96,7 @@
   booktitle = {Machine Learning and Knowledge Discovery in Databases},
   author = {Mohr, Felix and Viering, Tom J. and Loog, Marco and {\noopsort{rijn}}{van Rijn}, Jan N.},
   editor = {Amini, Massih-Reza and Canu, St{\'e}phane and Fischer, Asja and Guns, Tias and Kralj Novak, Petra and Tsoumakas, Grigorios},
+  doi = {10.1007/978-3-031-26419-1_1},
   year = {2023},
   pages = {3--19},
   publisher = {Springer Nature Switzerland},
