### 2026-01-05 1313077
recent publication addition to README
