+++ README.md
@@ -88,9 +88,13 @@ Any contributions are welcome. Please make sure you test what your contributions
 
 ## Project History and Development
 
-This work has been developed since 2023 and was previously used in [Turan et al. (2025)](#turan2025) under the name *mlcxx*. The name has since changed from *mlcxx* to *lcpp*. While the majority of the code remains the same, improvements in usability have been made, and unused parts of the code have been removed.  
+This work has been developed since 2023 and was previously used in [Turan et al. (2025)](#turan2025) under the name *mlcxx*. The name has since changed from *mlcxx* to *lcpp*. While the majority of the code remains the same, improvements in usability have been made, and unused parts of the code have been removed. It has most recently been used in [Turan et al. (2026)](#turan2026).
 
 ### Reference
 <a id="turan2025"></a>  
 Turan, O. T., Tax, D. M. J., Viering, T. J., & Loog, M. (2025). *Learning learning curves*. Pattern Analysis and Applications, 28, 15. [https://doi.org/10.1007/s10044-024-01394-6](https://doi.org/10.1007/s10044-024-01394-6)
+
+<a id="turan2026"></a>
+Turan, O. T., Loog, M., & Tax, D. M. J. (2026). *Generalization Performance Distributions Along Learning Curves*. Pattern Recognition Letters. [https://doi.org/10.1016/j.patrec.2026.01.003] (https://doi.org/10.1016/j.patrec.2026.01.003)
+
 
