+++ docs/paper/ref.bib
@@ -132,9 +132,14 @@ address = {New York, NY, USA},
 }
 
 
-@misc{turan2025b,
-  author    = {Turan, O. Taylan and Loog, Marco and Tax, David M. J.},
-  title     = {Generalization Performance Distributions Along Learning Curves},
-  year      = {2025},
-  note      = {Manuscript in submission process}
+@article{turan2026,
+title = {Generalization Performance Distributions Along Learning Curves},
+journal = {Pattern Recognition Letters},
+year = {2026},
+issn = {0167-8655},
+doi = {https://doi.org/10.1016/j.patrec.2026.01.003},
+url = {https://www.sciencedirect.com/science/article/pii/S0167865526000012},
+author = {O. Taylan Turan and Marco Loog and David M J Tax},
+keywords = {learning curve, generalization performance, model selection, hyper-parameter tuning, quantiles},
+abstract = {Learning curves show the expected performance with respect to training set size. This is often used to evaluate and compare models, tune hyper-parameters and determine how much data is needed for a specific performance. However, the distributional properties of performance are frequently overlooked on learning curves. Generally, only an average with standard error or standard deviation is used. In this paper, we analyze the distributions of generalization performance on the learning curves. We compile a high-fidelity learning curve database, both with respect to training set size and repetitions of the sampling for a fixed training set size. Our investigation reveals that generalization performance rarely follows a Gaussian distribution for classical classifiers, regardless of dataset balance, loss function, sampling method, or hyper-parameter tuning along learning curves. Furthermore, we show that the choice of statistical summary, mean versus measures like quantiles affect the top model rankings. Our findings highlight the importance of considering different statistical measures and use of non-parametric approaches when evaluating and selecting machine learning models with learning curves.}
 }
