### 2026-01-08 #9737 @editorialbot
[REVIEW]: Learning Curve Plus Plus
**Submitting author:** @taylanot (<a href="http://orcid.org/0009-0004-1022-6334">Taylan Turan</a>)
**Repository:** https://github.com/taylanot/lcpp.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
