### 2026-01-08 #9737 @taylanot
@jolars When the work was submitted three months ago, I believe that a Software Design section, a Research Impact Statement, and an AI usage disclosure were not part of the standard requirements. Please let me know whether these need to be added now.
