+++ docs/paper/ref.bib
@@ -137,7 +137,7 @@ title = {Generalization Performance Distributions Along Learning Curves},
 journal = {Pattern Recognition Letters},
 year = {2026},
 issn = {0167-8655},
-doi = {https://doi.org/10.1016/j.patrec.2026.01.003},
+doi = {10.1016/j.patrec.2026.01.003},
 url = {https://www.sciencedirect.com/science/article/pii/S0167865526000012},
 author = {O. Taylan Turan and Marco Loog and David M J Tax},
 keywords = {learning curve, generalization performance, model selection, hyper-parameter tuning, quantiles},
