### 2026-01-08 #9737 @jolars
Yes, sorry, I should've mentioned this earlier. JOSS just undertook a scope change, and this review will be reviewed under the new scope and requirements. If you haven't yet, please read [the announcement](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future) and the [updated requirements](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance).

So yes, please do add these sections to your paper and then use `@editorialbot generate pdf` to regenerate the paper.

I'm sorry for the inconvenience this causes you.
