### 2026-01-23 #9737 @jolars
@zoq and @robcaulk - I just realized that editorialbot had generated the old version of the checklist for you. I have updated both of your checklists manually just now. The only substantial difference is that there is an **AI Usage** item now as well. I'm sorry about the mishap!
