### 2026-01-26 #9737 @rcurtin
Ok, here are the comments from my review.  Thanks @taylanot for the submission, I enjoyed reading it and playing with the software.

*Overall*: I think this is a good contribution, and it appears to be well-written software that is flexible enough to be deployed to a wide variety of problems.  There are some minor concerns about scope and related work that I detail below, but I'll defer to @jolars on their impact to the submission.  I was able to build the examples and tests and run them without major issues and I was able to gather enough information from the documentation to see how I might assemble my own learning curves.  Minor points below:

 * For ensmallen there is a better JMLR citation, if you don't mind, please use that one instead of the workshop version: https://www.jmlr.org/papers/v22/20-416.html

 * There is a typo in `data.md` in the documentation: `data::Dataset<arma::<Mat<double>> regdataset2(3)` (a stray `<`)

 * Just as an FYI, recently the mlpack developers (myself included) added a file [`mlpack.cmake`](https://github.com/mlpack/mlpack/blob/master/CMake/mlpack.cmake) which can be used to automatically download dependencies of mlpack.  This helps work around the confusing-for-users situation where the versions of mlpack in older Ubuntu LTS versions are out of date ... (I have tried for years to get anyone to pay attention to updating that, to no avail).  However, lcpp doesn't use CMake as a build system, so it may not be applicable.

 * The paper does not document any related efforts; without lcpp, what are my alternatives?  I can definitely believe that in C++-land, it's lcpp or hand-roll the code yourself, so it is a clear improvement there.  But how does this relate to tools in Python-land that surely exist for PyTorch and TensorFlow (or even for scikit)?  I think that C++ has enough benefits to justify the use of lcpp vs. other tools (for instance, mlpack gets a lot of traction in embedded and low-resource environments because it is lightweight header-only C++), but I still think it would be important to draw connections between lcpp and similar software.

 * The open development requirement suggests six months of effort, but technically the first commit here is August, giving only five months.  However, I infer that the original codebase may be from some time before that, as the two papers using lcpp date from earlier.

 * There is only one named contributor in the repository and commit history.  So this does not appear to be a collaborative effort, but the papers using lcpp have other authors.

 * The `ann_class.cpp` example does not compile for me:

```
examples/ann_class.cpp:41:30:   required from here
   41 |   network.Add<mlpack::Linear>(trainset.num_class_);
      |   ~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~
/usr/include/mlpack/methods/ann/layer/multi_layer.hpp:169:23: error: no matching function for call to 'mlpack::LinearType<arma::Mat<double>, mlpack::NoRegularizer>::LinearType(std::optional<long unsigned int>&)'
  169 |     network.push_back(new LayerType(args...));
```

easily fixed by using `trainset.num_class_.value_or(2)` (or similar).

 * I think that the documentation could benefit from a clearer "flow" and thinking about how a user would actually navigate the documentation.  I had a slightly difficult time figuring out where to go or where to start from; a step by step tutorial could be useful to this end, instead of just adaptable examples.  Or, perhaps just better organization of the documentation itself could solve this.
