### 2026-01-28 ee70f3b
fixing readme missing *run* statement
