### 2026-01-28 c9262a3
ann_class example bug fix
