### 2026-01-28 d6bf6e7
remove the stray < in data.md
