### 2026-01-29 b445ab8
rcurtin paper feedback addressed
