+++ docs/paper/paper.md
@@ -31,25 +31,31 @@ A learning algorithm is said to learn if its performance on a given task improve
 
 Learning curve databases are valuable for **model selection** and for **estimating the amount of data needed** to achieve a target performance. These applications typically assume learning curves are monotonic and convex. However, findings of [@yan2025], [@mohr2023] and [@viering2022] suggest that learning curves often exhibit more complex and irregular behavior. Sparse sampling of training sizes limits the ability to fully characterize these behaviors, highlighting the need for **high-fidelity learning curves** to investigate them.
 
-**LCPP (Learning Curve Plus Plus)** is a C++ library that allows for learning curve creation of machine learning models. LCPP enables its users to obtain learning curves for variety of learners on any supervised learning dataset with/out hyper-parameter tuning, enabling  model selection and training data requirement determination.
+**LCPP (Learning Curve Plus Plus)** is a C++ library that allows for learning curve creation of machine learning models. `LCPP` enables its users to obtain learning curves for variety of learners on any supervised learning dataset with/out hyper-parameter tuning, enabling  model selection and training data requirement determination.
 
 # Statement of need
 
 Generally, creating learning curves is computationally expensive because it requires repeatedly training algorithms on many subsets of varying training sizes. Consequently, learning curves are often computed for a limited number of training set sizes. For example, while creating learning curve databases [@mohr2023] and [@yan2025] limited number of training set-sizes are investigated, moreover, these generations are done only for fixed learners without hyper-parameter tuning.
 
-To empower the machine learning community to generate richer, more detailed learning curves, we propose LCPP, a C++ library for scalable learning curve generation. LCPP offers several features; first several approaches for splitting a given dataset into training and test sets of varying sizes (where training sets can be drawn randomly or incrementally, where test sets can be fixed or vary in size). Next, unlike most existing tools that fix hyper-parameters during learning curve creation, LCPP integrates hyperparameter optimization routines from mlpack [@mlpack2023], enabling optimized learner evaluations. 
+To empower the machine learning community to generate richer, more detailed learning curves, we propose `LCPP`, a C++ library for scalable learning curve generation. `LCPP` offers several features; first several approaches for splitting a given dataset into training and test sets of varying sizes (where training sets can be drawn randomly or incrementally, where test sets can be fixed or vary in size). Next, unlike most existing tools that fix hyper-parameters during learning curve creation, `LCPP` integrates hyperparameter optimization routines from `mlpack` [@mlpack2023], enabling optimized learner evaluations. 
 
-LCPP also includes a simple dataset container for access to OpenML datasets [@openml], with built-in support for complete dataset transformations and train/test splits, allowing users to directly measure the generalization performance of models available in mlpack [@mlpack2023] and some other learning algorithms included in itself, such as kernel ridge regression, discriminant classifiers, multi-class classification extensions of binary classifiers. 
+`LCPP` also includes a simple dataset container for access to `OpenML` datasets [@openml], with built-in support for complete dataset transformations and train/test splits, allowing users to directly measure the generalization performance of models available in `mlpack` [@mlpack2023] and some other learning algorithms included in itself, such as kernel ridge regression, discriminant classifiers, multi-class classification extensions of binary classifiers. 
+
+# State of the field 
+
[…]
