+++ docs/paper/ref.bib
@@ -24,18 +24,15 @@
 }
 
 
-@article{ensmallen2018,
-  title = {Ensmallen: A Flexible {{C}}++ Library for Efficient Function Optimization},
-  shorttitle = {Ensmallen},
-  author = {Bhardwaj, Shikhar and Curtin, Ryan R. and Edel, Marcus and Mentekidis, Yannis and Sanderson, Conrad},
-  year = {2018},
-  month = dec,
-  eprint = {1810.09361},
-  primaryclass = {cs, math},
-  doi = {10.5281/zenodo.2008650},
-  urldate = {2024-02-21},
-  archiveprefix = {arXiv},
-  keywords = {65K10 68N99 68W99 90C53,Computer Science - Machine Learning,Computer Science - Mathematical Software,G.1.6 G.4,Mathematics - Optimization and Control},
+@article{ensmallen2021,
+  author  = {Ryan R. Curtin and Marcus Edel and Rahul Ganesh Prabhu and Suryoday Basak and Zhihao Lou and Conrad Sanderson},
+  title   = {The ensmallen library for flexible numerical optimization},
+  journal = {Journal of Machine Learning Research},
+  year    = {2021},
+  volume  = {22},
+  number  = {166},
+  pages   = {1--6},
+  url     = {http://jmlr.org/papers/v22/20-416.html}
 }
 
 @article{mlpack2023,
