### 2026-01-29 #9737 @taylanot
@rcurtin Thanks for your detailed feedback and suggestions. They are valuable additions.

- I addressed the bugs, typos, and the citation.
- I made additions to the paper regarding the current learning curve generation tools.
- I also encountered some problems with `mlpack` on older `Ubuntu LTS` versions at certain times. I added your suggestion to the future contributions section. I intend to migrate to `CMake` in future versions.
- Regarding the commit history, yes, that is correct. The code-base evolved gradually from work that began in 2023, driven by practical needs and ongoing research, before it was formalized into this project. Due to concerns about unintended information leakage, I rebased the repository to its current state. As a result, the public commit history does not fully represent the complete development timeline.
