### 2026-01-29 058e5c6
ref.bib scikit fix
