### 2026-02-03 #9737 @jolars
Thanks for your detailed review, @rcurtin. Regarding the collaborative effort criterion, we have the following caveat in the review criteria:

> Single-author projects may be acceptable if they show other evidence of community engagement (extensive issues/discussions, external feature requests, documented users).

It is of course up to you to decide whether you believe this project meets that criteria or not. As I mentioned earlier, I did not find the submission to be out of scope, but we weigh the opinions of the reviewers heavily and will if necessary initialize a scope review if there is concern among the reviewers.
