### 2026-02-10 #9737 @zoq
Thanks [@taylanot](https://github.com/taylanot) for the submission, really interesting project, that follows a really thorough architecture. Before I lose my notes, I’ll start the documentation; Some suggestions for the README:

> You can also use the dockerfile to build your own image with docker or podman. Just run to build the image:

I would use Dockerfile, since it’s what is commonly used and it’s also the filename in the repository.

> After cloning this repository and runing ./install.sh, lcpp is at your disposal.

This should be running.

> Please make sure you test what your contributions in the related test files.

This sounds strange, but removing the what would fix it.
