### 2026-02-10 #9737 @robcaulk
For this submission, there seem to be some problems that I'd need JOSS editors to comment on:

1. The development group underpinning this project appears to be non-existent. I understand that the author rebased this and lost git history at some point, but there is no issue activity, no PR activity, and no commits from anyone besides the author. This seems to violate the "Contribution and Authorship" and nearly all the "Development history and open-source practice" points. There is no evidence of external engagement, and no evidence that that will start.

2. This has not been used in any scientific publications besides the original author's two publications. This seems to violate the "Scope and Significance" points of the submission guidelines.

3. The open-source practices are severely lacking here - no automated tests (barely manual tests without clear documentation), no github workflows, and no documentation besides a README. This does not meet JOSS standards.

For all these reasons, this project is not poised to attract and maintain external collaborators.

@jolars are you and the JOSS editorial team willing to see past these issues?
