### 2026-02-11 f5427a2
return of the test.cpp is ensured to return the result of the test
