### 2026-02-11 #9737 @taylanot
@jolars - I would like to provide some additional context regarding some points.

**Contribution and authorship:** I believe contributions to a software project can take many forms. While code commits are the most visible form, other contributions, such as driving discussions that lead to new features, or a simple research question can lead to a feature. These should be equally valuable in my opinion. In this project, several discussions and design decisions of multiple people played an important role in shaping the software.

**External engagement:** learning curve research remains a relatively small area, with relatively limited efforts compared to mainstream machine and deep learning topics. As a result, there is limited existing software that is verified, tested and published. One of the main motivations for trying to publish this work is to increase its visibility and adoption. While prior external engagement may be limited, we are optimistic about future interest. The potential for use within the broader `mlpack` community is notable, given its size and active user base. The publication efforts of our work aims to support wider adoption and engagement. Moreover, as Bandicoot adoption expands, LCPP has the potential to develop within that ecosystem.

**Github-Workflows:** I appreciate the observations regarding the missing testing documentation. I have now added this documentation and also included a GitHub workflow for automated testing.
