### 2026-02-13 #9737 @arfon
Hi folks, EiC of JOSS here. Thanks for all of the detailed feedback from the various reviews here. @jolars asked me to take a look at this thread, and in particular the issues flagged by @robcaulk [here](https://github.com/openjournals/joss-reviews/issues/9737#issuecomment-3879682103).

This submission should have been labeled as `pre-2026-submission` when we rolled out our updated scope and review criteria in January, but we missed it. That means this paper should be reviewed under our previous criteria, which were less prescriptive on some of the points raised.

To be clear: you don't need to update your review checklists to the old format, but if you'd like to see what the old review criteria looked like, they're available here: https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html

While the concerns raised may well be relevant under our new scope, they are not things we would have necessarily flagged under the previous criteria. I'm comfortable with the review proceeding on that basis.

Apologies for the confusion, and thanks again for your careful reviews.
