### 2026-02-17 #9737 @jolars
:wave: @robcaulk, @rcurtin, @zoq - please see the comment from arfon earlier here. I hope this resolves the questions regarding scope, please let me know if you have any remaining concerns regarding this point.
