### 2026-02-17 #9737 @robcaulk
Hello @jolars , my review is complete! My opinion is that this project warrants Major Revisions Required, it is still not even close to satisfying the pre-2026 criteria. In particular: https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html#substantial-scholarly-effort is missing all bullet points.

Documentation is also still quite poor.
