### 2026-02-18 #9737 @taylanot
Thanks for taking the time @robcaulk . @jolars , I understand that most of the issues raised are not actionable. Could you please outline the next steps in the procedure and clarify how we should proceed from here?
