### 2026-02-19 #9737 @jolars
@robcaulk - I understand, thanks for the clarification! Documentation is something that can be addressed in the review process, but other issues are less actionable, as @taylanot also mentioned. @robcaulk, is there any chance you might be able to comment a bit more specifically on which parts of the pre-2026 criteria you think are not fulfilled and why? I am, for instance, interested in hearing what you have to say on "whether the software is of broad interest (likely to be cited by other researchers)"?
