### 2026-02-19 36c2df2
add doxygen workflow
