### 2026-02-19 #9737 @taylanot
@jolars  – I have added an automated documentation generation workflow to make the project easier to understand, more accessible, and more sustainable. I would welcome any suggestions for further improvements or refinements.
