### 2026-02-27 #9737 @robcaulk
Hey everyone,

I'd like to point out that if y'all are happy with this, you do not need my approval to move forward. All I am saying is that this is, in my opinion, pretty clearly violating some of the most basic requirements of a JOSS endorsed project. My opinion may easily be wrong - happy to let you all move forward against it if you disagree.

@jolars in response to your ping:

> Age of software (is this a well-established software project) / length of commit history.

This repo has approximately 6 months of commit history. From a single person. This does not to me, indicate a well-established software project.

At risk of repeating earlier parts of this thread: there are no issues, no forks, no PRs. These are pretty clear indications that this project is not a well-established software project.

> Number of authors.

One committer. This does not to me, indicate a well established software project.

> Lines of code (LOC): These statistics are usually reported by EditorialBot in the pre-review issue thread.

Not sure what the point of this criteria is.

> Whether the software has already been cited in academic papers.

It has not been cited by anyone except the author himself. Two times. This does not to me, indicate a significant project with any academic traction.

> Whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain.

This is a speculative question. Usually the only way to avoid pure speculation is to look at who has cited it so far. We would look to see if anyone has cited this, and in what capacity have they cited it. Since this has no citations, there is no way to know.
