### 2026-02-27 #9737 @rcurtin
I don't disagree with anything Rob wrote (I mean it's not possible to disagree with those observations).  As a very occasional JOSS reviewer I don't know if those observations are disqualifying for publication, so I can't comment there, but I can at least add a word to indicate that each factual observation made in the response above is correct.
