### 2026-03-04 #9737 @taylanot
Thank you all for the continued discussion.

First, I would like to note that the concerns regarding the commit history were already mentioned during the initial submission phase by me. In other words, this aspect was known prior to the start of the review process. For that reason, revisiting the same topic after six months of discussion feels somewhat unproductive. Nevertheless, I would still like to clarify a few points from my perspective.

### Age of the software

The first paper that cites our work under the name *mlcxx*, *Learning learning curves*, was received by PAA journal in August 2024. I believe this serves as a proxy indicating that the software is older than six months. Since the *mlcxx* link in that paper directs to the *lcpp* repository, this also provides an observational indication that the repository existed at least a year before the submission to JOSS.

### Already cited in scientific work

My understanding of this criterion is that it aims to determine whether the software has a genuine scientific use case, whether it is extensible for research purposes, and whether it targets research use. The guideline does not explicitly distinguish between self citations and citations by independent work. Therefore, I interpret the question more broadly: does the software enable scientific output?

### Likelihood of future citations

I agree that predicting future citations is inherently speculative. However, I do not believe that relying solely on past citation counts, as Robert suggests, provides a reliable way to estimate future impact. The guidelines appear to suggest an informed judgment, instead of looking at past citations. In my view, such judgment should consider the current needs and direction of the research field. For example, it may be more informative to consider questions such as which problems the software addresses, whether it fills an existing gap, whether it is likely to remain useful in the near future, and whether meaningful alternatives exist.
