### 2026-03-24 #9737 @crvernon
Hello, I am the associate editor-in-chief of this track of submissions.  First, thank you @robcaulk for the [expansion of your comments](https://github.com/openjournals/joss-reviews/issues/9737#issuecomment-3973233317) as requested by @jolars.  Your points are much appreciated and valid if this submission were to be viewed under our new criteria. These types of considerations, among others, were what prompted some of our new policy for substantial scholarly effort now in effect.  However, since this submission qualifies under our pre-2026 criteria as [Arfon pointed out](https://github.com/openjournals/joss-reviews/issues/9737#issuecomment-3897076147), I believe we are good to move forward with this review as-is permitting the approval of your topic editor @jolars to satisfy two reviews in full.  Again, thank you for your feedback and @taylanot the responses and clarification from your side.
