### 2026-03-25 #9737 @robcaulk
Thanks @crvernon , just for the record, this comment is based entirely on pre-2026 criteria: https://github.com/openjournals/joss-reviews/issues/9737#issuecomment-3973233317

Glad you found a way to move forward!
