### 2026-03-27 #9737 @jolars
Thanks for everyone's patience, and thanks for the clarification @crvernon!

:wave: @zoq, @rcurtin - I am pasting the pre-2026 review checklist below. Please say if you do or do not think the current submission fulfills these criteria. (I understand that you do not, @robcaulk.)

## Review checklist for NAME

- [ ] **Repository:** Is the source code for this software available at [https://github.com/taylanot/lcpp.git](https://github.com/taylanot/lcpp.git)?
