### 2026-04-01 #9737 @rcurtin
Sorry!  I actually had not noticed that that required my reply (it very obviously does, that is totally on me).  Since the pre-2026 checklist drops the 'scope and significance' and 'collaborative effort' guidelines, then at least from my perspective the work satisfies all criteria.
