### 2026-04-07 #9737 @zoq
Same as @rcurtin, if 'scope and significance' is not part of the checklist it covers everything; if you include it it lacks on that part as pointed out before.
