### 2026-04-07 #9737 @robcaulk
If "scope and significance" and "collaborative effort" are actually not in the pre-2026 checklist, @editors might want to remove them from the pre-2026 checklist: https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html#substantial-scholarly-effort

That is what I based my entire criticism on.
