### 2026-04-07 #9737 @jolars
Thanks for the clarifications, @zoq and @rcurtin!

> If "scope and significance" and "collaborative effort" are actually not in the pre-2026 checklist, @editors might want to remove them from the pre-2026 checklist: [joss.readthedocs.io/en/pre-2026-submission/review_criteria.html#substantial-scholarly-effort](https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html#substantial-scholarly-effort)

Thanks for the input, @robcaulk, but I don't believe the pre-2026 checklist includes "collaborative effort" as part of the review criteria. Scope and significance is, however, part of the "substantial scholary effort" item that you linked to. But note that the list there is not meant to be strictly prescriptive.
