### 2026-04-12 #9737 @taylanot
Hello @jolars, thanks for the checks and also suggestions to the paper.

- I checked the title again for special characters.
- Created a new release that incorporates the changes during the review.
- Archived this release in Zenodo with the same licence.

- DOI: 10.5281/zenodo.19534268
-  Version tag that is released: v1.1
