### 2026-04-22 #9737 @crvernon
🥳 Congratulations on your new publication @taylanot ! Many thanks to @jolars for editing and @rcurtin, @zoq, and @robcaulk for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
